consolidates GameUiContext and TeleporterContext
