+++ SolastaCommunityExpansion/Patches/GameUiLocation/GameLocationScreenExplorationPatcher.cs
@@ -14,32 +14,41 @@ namespace SolastaCommunityExpansion.Patches.GameUiLocation
             CharacterControlPanelExploration ___characterControlPanelExploration,
             TimeAndNavigationPanel ___timeAndNavigationPanel)
         {
-            if (!Main.Settings.EnableHotkeysToToggleHud)
+            if (Main.Settings.EnableHotkeysToToggleHud)
             {
-                return;
+                switch (command)
+                {
+                    case Hotkeys.CTRL_C:
+                        ShowCharacterControlPanelExploration();
+                        break;
+
+                    case Hotkeys.CTRL_L:
+                        ShowGuiConsoleScreen();
+                        break;
+
+                    case Hotkeys.CTRL_M:
+                        ShowTimeAndNavigationPanel();
+                        break;
+
+                    case Hotkeys.CTRL_P:
+                        ShowPartyControlPanel();
+                        break;
+
+                    case Hotkeys.CTRL_H:
+                        ServiceRepository.GetService<IDebugOverlayService>()?.ToggleActivation();
+                        ShowAll();
+                        break;
+                }
             }
 
-            switch (command)
+            if (Main.Settings.EnableDebugOverlay && command == Hotkeys.CTRL_SHIFT_D)
             {
-                case Hotkeys.CTRL_C:
-                    ShowCharacterControlPanelExploration();
-                    break;
+                ServiceRepository.GetService<IDebugOverlayService>()?.ToggleActivation();
+            }
 
-                case Hotkeys.CTRL_L:
-                    ShowGuiConsoleScreen();
-                    break;
-
-                case Hotkeys.CTRL_M:
-                    ShowTimeAndNavigationPanel();
-                    break;
-
-                case Hotkeys.CTRL_P:
-                    ShowPartyControlPanel();
-                    break;
-
-                case Hotkeys.CTRL_H:
-                    ShowAll();
-                    break;
+            if (Main.Settings.EnableTeleportParty && command == Hotkeys.CTRL_SHIFT_T)
+            {
+                Models.GameUiContext.Teleporter.ConfirmTeleportParty();
             }
 
             void ShowAll()
